requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
coverage==6.2
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.22.1
pandas==1.3.5
phonenumberslite==8.12.41
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
